CET-308: 500 error when a workflow creates a new related meeting (#33219)

+++ sugarcrm/data/SugarBean.php
@@ -2498,18 +2498,22 @@ class SugarBean
                     }
                     $GLOBALS['log']->debug('save_relationship_changes(): From relationship_field array - adding a relationship record: '.$rel_name . ' = ' . $this->$id);
                     //already related the new relationship id so let's set it to false so we don't add it again using the _REQUEST['relate_i'] mechanism in a later block
-                    $this->load_relationship($rel_name);
-                    $rel_add = $this->$rel_name->add($this->$id);
-                    // move this around to only take out the id if it was save successfully
-                    if ($this->$id == $new_rel_id && $rel_add == true) {
-                        $new_rel_id = false;
+                    //ut exempt to be used with unit tests that mock link classes
+                    if ($this->load_relationship($rel_name)) {
+                        $rel_add = $this->$rel_name->add($this->$id);
+                        // move this around to only take out the id if it was save successfully
+                        if ($this->$id == $new_rel_id && $rel_add == true) {
+                            $new_rel_id = false;
+                        }
                     }
                 } else {
                     //if before value is not empty then attempt to delete relationship
                     if (!empty($this->rel_fields_before_value[$id])) {
                         $GLOBALS['log']->debug('save_relationship_changes(): From relationship_field array - attempting to remove the relationship record, using relationship attribute' . $rel_name);
-                        $this->load_relationship($rel_name);
-                        $this->$rel_name->delete($this->id, $this->rel_fields_before_value[$id]);
+                        //ut exempt to be used with unit tests that mock link classes
+                        if ($this->load_relationship($rel_name)) {
+                            $this->$rel_name->delete($this->id, $this->rel_fields_before_value[$id]);
+                        }
                     }
                 }
             }
